change: replace to new SideBar
